refactor(bridge): separate browser and capture handlers
