+++ Core/PeekabooCore/Sources/PeekabooBridge/PeekabooBridgeServer+Handlers.swift
@@ -3,7 +3,6 @@ import Foundation
 @_spi(Bridge) import PeekabooAutomationKit
 import PeekabooFoundation
 
-// swiftlint:disable file_length
 @MainActor
 extension PeekabooBridgeServer {
     // swiftlint:disable:next cyclomatic_complexity
@@ -130,362 +129,6 @@ extension PeekabooBridgeServer {
         }
     }
 
-    private func handleBrowserRequest(_ request: PeekabooBridgeRequest) async throws
-        -> PeekabooBridgeResponse
-    {
-        switch request {
-        case let .browserStatus(payload):
-            return try await .browserStatus(self.services.browserStatus(channel: payload.channel))
-        case let .browserConnect(payload):
-            let status = try await self.legacyBrowserConnectionStatus(payload)
-            return .browserStatus(status)
-        case .browserDisconnect:
-            try await self.services.browserDisconnect()
-            return .ok
-        case let .browserExecute(payload):
-            return try await .browserToolResponse(self.services.browserExecute(payload))
-        default:
-            throw Self.invalidRequest(for: request)
-        }
-    }
-
-    private func handleScopedBrowserStatus(
-        sessionID: UUID,
-        channel: String?) async throws -> PeekabooBridgeBrowserStatus
-    {
-        let caller = try self.authenticatedBrowserSessionCaller()
-        let lease = try await self.browserHandoffGrantRegistry.authorizeSession(sessionID, caller: caller)
-        defer { self.browserHandoffGrantRegistry.completeSessionOperation(lease) }
-        guard let provider = self.browserSessionBootstrapProvider else {
-            throw PeekabooBridgeErrorEnvelope(
-                code: .operationNotSupported,
-                message: "This Bridge host has no scoped browser session provider")
-        }
-        let status = try await provider.browserSessionStatus(sessionID: sessionID, channel: channel)
-        guard status.isCanonicalScopedSessionStatus else {
-            throw PeekabooBridgeErrorEnvelope(
-                code: .internalError,
-                message: "Scoped browser provider returned contradictory status evidence")
-        }
-        return status
-    }
-
-    private func handleBrowserSessionControl(
-        _ request: PeekabooBridgeBrowserSessionControlRequest) async throws -> PeekabooBridgeResponse
-    {
-        let caller = try self.authenticatedBrowserSessionCaller()
-        guard let provider = self.browserSessionBootstrapProvider else {
-            throw PeekabooBridgeErrorEnvelope(
-                code: .operationNotSupported,
-                message: "This Bridge host has no scoped browser session provider")
-        }
-        switch request.action {
-        case .disconnect:
-            let lease = try await self.browserHandoffGrantRegistry.authorizeSession(
-                request.sessionID,
-                caller: caller)
-            defer { self.browserHandoffGrantRegistry.completeSessionOperation(lease) }
-            try await provider.disconnectBrowserSession(request.sessionID)
-        case .end:
-            try await self.browserHandoffGrantRegistry.endSession(request.sessionID, caller: caller)
-        }
-        return .ok
-    }
-
-    func authenticatedBrowserSessionCaller() throws -> PeekabooBridgeBrowserSessionCaller {
-        guard let operation = PeekabooBridgeRequestContext.browserHandoffOperation else {
-            throw PeekabooBridgeErrorEnvelope(
-                code: .unauthorizedClient,
-                message: "Scoped browser operations require an attested operation caller")
-        }
-        return try operation.peer.browserSessionCaller(clientInstanceID: operation.clientInstanceID)
-    }
-
-    private func handleBrowserExecute(
-        _ payload: PeekabooBridgeBrowserExecuteRequest) async throws -> PeekabooBridgeHandledResponse
-    {
-        guard !payload.resolvedCalls.isEmpty else {
-            throw DesktopActionFailure.preDispatchRefusal(
-                reason: .invalidRequest,
-                message: "Browser execution requires at least one tool call.",
-                hint: "Provide one browser tool call before retrying.")
-        }
-        let target: (
-            receipt: PeekabooBridgeBrowserConnectionReceipt,
-            disposition: PeekabooBridgeHandledResponse.Mutation.TargetDisposition)
-        let scopedSessionID = payload.sessionID
-        var scopedLease: PeekabooBridgeBrowserHandoffGrantRegistry.SessionOperationLease?
-        defer {
-            if let scopedLease {
-                self.browserHandoffGrantRegistry.completeSessionOperation(scopedLease)
-            }
-        }
-        if let scopedSessionID {
-            let caller = try self.authenticatedBrowserSessionCaller()
-            scopedLease = try await self.browserHandoffGrantRegistry.authorizeSession(
-                scopedSessionID,
-                caller: caller)
-            guard let expectedReceipt = payload.expectedConnectionReceipt,
-                  expectedReceipt.isCanonicalExecutionTarget,
-                  payload.expectedProviderSessionEpoch != nil,
-                  payload.connectionPolicy == .requireExistingLiveReceipt,
-                  payload.elementPreflight?.isCanonical != false
-            else {
-                throw DesktopActionFailure.preDispatchRefusal(
-                    reason: .invalidRequest,
-                    message: "Scoped browser execution requires an exact receipt, epoch, and canonical preflight.",
-                    hint: "Refresh scoped browser status before retrying.")
-            }
-            target = try (expectedReceipt, self.browserTargetDisposition(expectedReceipt))
-        } else {
-            scopedLease = nil
-            target = try await self.browserExecutionTarget(payload)
-        }
-        let result: PeekabooBridgeBrowserExecutionResult
-        do {
-            if let scopedSessionID {
-                guard let provider = self.browserSessionBootstrapProvider else {
-                    throw PeekabooBridgeErrorEnvelope(
-                        code: .operationNotSupported,
-                        message: "This Bridge host has no scoped browser session provider")
-                }
-                result = try await provider.browserSessionExecute(
-                    sessionID: scopedSessionID,
-                    request: payload,
-                    expectedConnectionReceipt: target.receipt)
-            } else {
-                result = try await self.services.browserExecute(
-                    payload,
-                    expectedConnectionReceipt: target.receipt)
-            }
-        } catch is CancellationError {
-            let acceptedProviderSessionEpoch = scopedSessionID == nil
-                ? nil
-                : payload.expectedProviderSessionEpoch
-            if payload.isReadOnly {
-                return Self.browserReadCancellationHandledResponse(
-                    target: target,
-                    providerSessionEpoch: acceptedProviderSessionEpoch)
-            }
-            return Self.browserOpaqueCancellationHandledResponse(
-                target: target,
-                providerSessionEpoch: acceptedProviderSessionEpoch,
-                causeDescription:
-                "The browser provider was cancelled after accepting the execution request.")
-        }
-        guard result.connectionReceipt == target.receipt,
-              scopedSessionID == nil || result.providerSessionEpoch == payload.expectedProviderSessionEpoch
-        else {
-            if payload.isReadOnly {
-                throw DesktopActionFailure.preDispatchRefusal(
-                    reason: .targetUnavailable,
-                    message: "Browser read returned a different connection receipt.",
-                    hint: "Refresh browser status and retry against its exact connection receipt.")
-            }
-            throw DesktopActionFailure.indeterminate(
-                evidence: .completionUnknown,
-                message: "Browser execution returned a different connection receipt.",
-                hint: "Observe the intended browser before retrying and update the runtime host.")
-        }
-        if payload.isReadOnly {
-            return try Self.browserReadHandledResponse(result: result, request: payload, target: target)
-        }
-        guard
-            result.completedCallCount >= 0,
-            result.dispatchedCallCount >= result.completedCallCount,
-            result.dispatchedCallCount <= payload.mutationCallCount
-        else {
-            throw DesktopActionFailure.indeterminate(
-                delivery: .init(mechanism: .browserProtocol, mode: .background),
-                evidence: .completionUnknown,
-                message: "Browser execution returned a different connection receipt.",
-                hint: "Observe the intended browser before retrying and update the runtime host.")
-        }
-        if result.dispatchedCallCount == 0 {
-            return try Self.browserNoDispatchHandledResponse(result: result, target: target)
-        }
-        guard
-            let dispatchedCallCount = DesktopActionOutcome.DispatchUnitCount(result.dispatchedCallCount)
-        else {
-            preconditionFailure("A positive browser dispatch count must have a canonical unit count")
-        }
-        let routedFailure: DesktopActionFailure? =
-            if let failure = result.actionFailure,
-            failure.outcome.dispatchState.mutationDispatched,
-            failure.outcome.dispatchState.unitCount == dispatchedCallCount {
-                failure.routed(to: .bridge)
-            } else if result.actionFailure != nil || result.response.isError
-                || result.completedCallCount != payload.mutationCallCount
-            {
-                Self.browserProviderIndeterminateFailure(
-                    completedCallCount: result.completedCallCount,
-                    dispatchedCallCount: dispatchedCallCount,
-                    causeDescription:
-                    "The browser provider returned incomplete or contradictory result semantics.")
-            } else {
-                nil
-            }
-        let response = PeekabooBridgeBrowserToolResponse(
-            content: result.response.content,
-            isError: routedFailure != nil,
-            meta: result.response.meta,
-            structuredContent: result.response.structuredContent,
-            connectionReceipt: target.receipt,
-            completedCallCount: result.completedCallCount,
-            dispatchedCallCount: result.dispatchedCallCount,
-            actionFailure: routedFailure,
-            providerSessionEpoch: result.providerSessionEpoch)
-        let outcome =
-            routedFailure?.outcome
-                ?? .dispatchedUnverified(
-                    delivery: .init(mechanism: .browserProtocol, mode: .background),
-                    evidence: .deliveryAccepted,
-                    unitCount: dispatchedCallCount)
-        return .init(
-            response: .browserToolResponse(response),
-            mutation: .init(
-                outcome: outcome,
-                target: target.disposition))
-    }
-
-    private static func browserReadCancellationHandledResponse(
-        target: (
-            receipt: PeekabooBridgeBrowserConnectionReceipt,
-            disposition: PeekabooBridgeHandledResponse.Mutation.TargetDisposition),
-        providerSessionEpoch: UUID?)
-        -> PeekabooBridgeHandledResponse
-    {
-        .init(response: .browserToolResponse(.init(
-            content: [
-                .object([
-                    "type": .string("text"),
-                    "text": .string("Browser read was cancelled after provider entry."),
-                ]),
-            ],
-            isError: true,
-            meta: nil,
-            connectionReceipt: target.receipt,
-            providerSessionEpoch: providerSessionEpoch)))
-    }
-
-    private static func browserReadHandledResponse(
-        result: PeekabooBridgeBrowserExecutionResult,
-        request: PeekabooBridgeBrowserExecuteRequest,
-        target: (
-            receipt: PeekabooBridgeBrowserConnectionReceipt,
-            disposition: PeekabooBridgeHandledResponse.Mutation.TargetDisposition)) throws
-        -> PeekabooBridgeHandledResponse
-    {
-        let callCount = request.resolvedCalls.count
-        guard result.completedCallCount >= 0,
-              result.dispatchedCallCount >= result.completedCallCount,
-              result.completedCallCount <= callCount,
-              result.dispatchedCallCount <= callCount,
-              result.response.isError || (
-                  result.completedCallCount == callCount &&
-                      result.dispatchedCallCount == callCount &&
-                      result.actionFailure == nil)
-        else {
-            throw DesktopActionFailure.preDispatchRefusal(
-                reason: .targetUnavailable,
-                message: "Receipt-bound browser read did not complete canonically.",
-                hint: "Refresh browser status and retry against its exact connection receipt.")
-        }
-        return .init(response: .browserToolResponse(.init(
-            content: result.response.content,
-            isError: result.response.isError,
-            meta: result.response.meta,
-            structuredContent: result.response.structuredContent,
-            connectionReceipt: target.receipt,
-            completedCallCount: result.completedCallCount,
-            dispatchedCallCount: result.dispatchedCallCount,
-            providerSessionEpoch: result.providerSessionEpoch)))
-    }
-
-    private static func browserNoDispatchHandledResponse(
-        result: PeekabooBridgeBrowserExecutionResult,
-        target: (
-            receipt: PeekabooBridgeBrowserConnectionReceipt,
-            disposition: PeekabooBridgeHandledResponse.Mutation.TargetDisposition)) throws
-        -> PeekabooBridgeHandledResponse
-    {
-        guard result.completedCallCount == 0,
-              let failure = result.actionFailure,
-              failure.outcome.state == .refused,
-              failure.outcome.dispatchState == .none,
-              failure.outcome.retrySafety == .safe,
-              failure.outcome.refusalReason != nil
-        else {
-            throw DesktopActionFailure.indeterminate(
-                route: .bridge,
-                delivery: .init(mechanism: .browserProtocol, mode: .background),
-                evidence: .completionUnknown,
-                message: "Browser provider returned contradictory zero-progress semantics.",
-                hint: "Observe the exact browser before retrying and update the runtime host.")
-        }
-        let routedFailure = failure.routed(to: .bridge)
-        return .init(
-            response: .browserToolResponse(
-                .init(
-                    content: result.response.content,
-                    isError: true,
-                    meta: result.response.meta,
-                    structuredContent: result.response.structuredContent,
-                    connectionReceipt: target.receipt,
-                    completedCallCount: 0,
-                    dispatchedCallCount: 0,
-                    actionFailure: routedFailure,
-                    providerSessionEpoch: result.providerSessionEpoch)),
-            mutation: .init(outcome: routedFailure.outcome, target: target.disposition))
-    }
-
-    private static func browserOpaqueCancellationHandledResponse(
-        target: (
-            receipt: PeekabooBridgeBrowserConnectionReceipt,
-            disposition: PeekabooBridgeHandledResponse.Mutation.TargetDisposition),
-        providerSessionEpoch: UUID?,
-        causeDescription: String) -> PeekabooBridgeHandledResponse
-    {
-        let failure = DesktopActionFailure.indeterminate(
-            route: .bridge,
-            delivery: .init(mechanism: .browserProtocol, mode: .background),
-            evidence: .completionUnknown,
-            message: "Browser execution completion is indeterminate; exact progress is unavailable.",
-            hint: "Observe the exact browser before deciding which work remains unfinished.",
-            causeDescription: causeDescription)
-        return .init(
-            response: .browserToolResponse(
-                .init(
-                    content: [
-                        .object([
-                            "type": .string("text"),
-                            "text": .string(failure.message),
-                        ]),
-                    ],
-                    isError: true,
-                    meta: nil,
-                    connectionReceipt: target.receipt,
-                    actionFailure: failure,
-                    providerSessionEpoch: providerSessionEpoch)),
-            mutation: .init(outcome: failure.outcome, target: target.disposition))
-    }
-
-    private static func browserProviderIndeterminateFailure(
-        completedCallCount: Int,
-        dispatchedCallCount: DesktopActionOutcome.DispatchUnitCount,
-        causeDescription: String) -> DesktopActionFailure
-    {
-        .indeterminate(
-            route: .bridge,
-            delivery: .init(mechanism: .browserProtocol, mode: .background),
-            evidence: .completionUnknown,
-            unitCount: dispatchedCallCount,
-            message: "Browser execution completion is indeterminate "
-                + "(\(completedCallCount) completed, \(dispatchedCallCount.rawValue) dispatched or accepted).",
-            hint: "Observe the exact browser before resuming unfinished work.",
-            causeDescription: causeDescription)
-    }
-
     private func handlePostEventPermissionRequest() -> PeekabooBridgeHandledResponse {
         let granted = self.postEventAccessRequester()
         return .init(
@@ -536,223 +179,6 @@ extension PeekabooBridgeServer {
         default:
             throw Self.invalidRequest(for: request)
         }
-    }
-
-    private func handleCaptureRequest(_ request: PeekabooBridgeRequest) async throws
-        -> PeekabooBridgeResponse
-    {
-        switch request {
-        case let .captureScreen(payload):
-            let capture = try await self.services.screenCapture.captureScreen(
-                displayIndex: payload.displayIndex,
-                visualizerMode: payload.visualizerMode,
-                scale: payload.scale)
-            return .capture(capture)
-        case let .captureWindow(payload):
-            return try await self.handleCaptureWindow(payload)
-        case let .captureFrontmost(payload):
-            let capture = try await self.services.screenCapture.captureFrontmost(
-                visualizerMode: payload.visualizerMode,
-                scale: payload.scale)
-            return .capture(capture)
-        case let .captureArea(payload):
-            let capture = try await self.services.screenCapture.captureArea(
-                payload.rect,
-                visualizerMode: payload.visualizerMode,
-                scale: payload.scale)
-            return .capture(capture)
-        default:
-            throw Self.invalidRequest(for: request)
-        }
-    }
-
-    private func handleCaptureWindow(
-        _ payload: PeekabooBridgeCaptureWindowRequest) async throws -> PeekabooBridgeResponse
-    {
-        if let windowID = try payload.validatedWindowID() {
-            let capture = try await self.services.screenCapture.captureWindow(
-                windowID: windowID,
-                visualizerMode: payload.visualizerMode,
-                scale: payload.scale)
-            return .capture(capture)
-        }
-
-        guard !payload.appIdentifier.isEmpty else {
-            throw PeekabooBridgeErrorEnvelope(
-                code: .invalidRequest,
-                message: "captureWindow requires appIdentifier or windowId")
-        }
-
-        let capture = try await self.services.screenCapture.captureWindow(
-            appIdentifier: payload.appIdentifier,
-            windowIndex: payload.windowIndex,
-            visualizerMode: payload.visualizerMode,
-            scale: payload.scale)
-        return .capture(capture)
-    }
-
-    private func handleDesktopObservationRequest(_ request: PeekabooBridgeRequest) async throws
-        -> PeekabooBridgeHandledResponse
-    {
-        switch request {
-        case let .desktopObservation(payload):
-            try Self.validateAttestedWebFocusTarget(payload)
-            let hostRegisteredScreenCaptureKitOwnership = self.hostCapabilities.contains(
-                PeekabooBridgeHostCapability.screenCaptureKitProcessOwnership)
-            let currentScreenCaptureKitOwnerReceipt: ScreenCaptureKitOwnerLease.OwnerReceipt? = if
-                hostRegisteredScreenCaptureKitOwnership,
-                payload.capture.engine == .auto,
-                payload.capture.focus == .background,
-                case .screen = payload.target
-            {
-                try? self.screenCaptureKitOwnerClaimProvider()
-            } else {
-                nil
-            }
-            let prefersModernFirstAutomaticCapture = Self.desktopObservationPrefersModernFirstAutomaticCapture(
-                payload,
-                hostRegisteredScreenCaptureKitOwnership: hostRegisteredScreenCaptureKitOwnership,
-                hostIdentity: self.hostIdentity,
-                currentScreenCaptureKitOwnerReceipt: currentScreenCaptureKitOwnerReceipt)
-            if self.services.desktopObservation is any DesktopObservationActionResultProviding {
-                let result = try await ScreenCaptureService.withModernFirstAutomaticCapture(
-                    prefersModernFirstAutomaticCapture)
-                {
-                    try await self.services.desktopObservation.observeResult(payload)
-                }
-                try Self.validateAttestedObservationBinding(
-                    payload,
-                    result: result.payload,
-                    requireContentDigest: false)
-                let attested = try result.payload.attestingCaptureContent()
-                try Self.validateAttestedObservationBinding(payload, result: attested)
-                let response = PeekabooBridgeResponse.desktopObservation(attested.withoutImageData())
-                guard request.mayMutateDesktop else {
-                    if let failure = Self.readOnlyObservationFailure(result) {
-                        let target = try PeekabooBridgeOperationTargetAttribution.resolve(
-                            request: request,
-                            response: response,
-                            handledTarget: result.targetIdentity)
-                        return .init(
-                            response: .error(
-                                .init(
-                                    code: .internalError,
-                                    actionFailure: failure.routed(to: .bridge))),
-                            targetIdentity: target)
-                    }
-                    return .init(response: response, targetIdentity: result.targetIdentity)
-                }
-                let outcome = try Self.requireSuccessfulObservationOutcome(result)
-                let target =
-                    result.targetIdentity.map {
-                        PeekabooBridgeHandledResponse.Mutation.TargetDisposition.handlerResolved($0)
-                    } ?? Self.observationFallbackTarget(for: payload)
-                if let failure = Self.observationFailure(result, outcome: outcome) {
-                    return .init(
-                        response: .error(
-                            .init(
-                                code: .internalError,
-                                actionFailure: failure.routed(to: .bridge))),
-                        mutation: .init(outcome: outcome, target: target))
-                }
-                return .init(
-                    response: response,
-                    mutation: .init(outcome: outcome, target: target))
-            }
-            if case let .menubarPopover(_, openIfNeeded) = payload.target,
-               openIfNeeded != nil
-            {
-                throw DesktopActionFailure.preDispatchRefusal(
-                    reason: .runtimeIncompatible,
-                    message: "Menu-bar popover opening requires an action-result-aware observation service.",
-                    hint: "Update the runtime host before retrying this conditional background mutation.")
-            }
-            let observation = try await ScreenCaptureService.withModernFirstAutomaticCapture(
-                prefersModernFirstAutomaticCapture)
-            {
-                try await self.services.desktopObservation.observe(payload)
-            }
-            try Self.validateAttestedObservationBinding(
-                payload,
-                result: observation,
-                requireContentDigest: false)
-            let attested = try observation.attestingCaptureContent()
-            try Self.validateAttestedObservationBinding(payload, result: attested)
-            let response = PeekabooBridgeResponse.desktopObservation(attested.withoutImageData())
-            guard request.mayMutateDesktop else {
-                return .init(response: response)
-            }
-            let mode: DesktopActionOutcome.Delivery.Mode =
-                payload.capture.focus == .background
-                    ? .background
-                    : .foreground
-            return .init(
-                response: response,
-                mutation: .init(
-                    outcome: .dispatchedUnverified(
-                        delivery: .init(mechanism: .capturePipeline, mode: mode),
-                        evidence: .deliveryAccepted,
-                        unitCount: .one),
-                    target: Self.observationFallbackTarget(for: payload)))
-        default:
-            throw Self.invalidRequest(for: request)
-        }
-    }
-
-    /// A Bridge-owned ScreenCaptureKit process can capture a background display directly. Keeping `auto` as
-    /// classic-first here serializes every request behind `/usr/sbin/screencapture`; that helper can stall despite the
-    /// host's usable ScreenCaptureKit grant and consume half of the Bridge deadline before modern fallback begins.
-    /// Registration and preparation do not claim the process-lifetime lease. The eligible request therefore claims
-    /// it atomically, then requires the returned live receipt to match this Bridge generation before selecting a
-    /// scoped modern-first automatic order. Legacy remains the fallback, and explicit engine choices plus
-    /// caller-local capture retain their existing contracts.
-    static func desktopObservationPrefersModernFirstAutomaticCapture(
-        _ request: DesktopObservationRequest,
-        hostRegisteredScreenCaptureKitOwnership: Bool,
-        hostIdentity: PeekabooBridgeHostIdentity?,
-        currentScreenCaptureKitOwnerReceipt: ScreenCaptureKitOwnerLease.OwnerReceipt?)
-        -> Bool
-    {
-        guard hostRegisteredScreenCaptureKitOwnership,
-              let hostIdentity,
-              let hostProcessStartIdentity = hostIdentity.processStartIdentity,
-              let currentScreenCaptureKitOwnerReceipt,
-              currentScreenCaptureKitOwnerReceipt.processIdentifier == hostIdentity.processIdentifier,
-              currentScreenCaptureKitOwnerReceipt.processStartIdentity == hostProcessStartIdentity,
-              request.capture.engine == .auto,
-              request.capture.focus == .background,
-              case .screen = request.target
-        else { return false }
-        return true
-    }
-
-    private static func readOnlyObservationFailure(
-        _ result: UIAutomationActionResult<DesktopObservationResult>) -> DesktopActionFailure?
-    {
-        guard let outcome = result.outcome else { return nil }
-        let targetReceipt = result.targetIdentity?.actionTargetReceipt
-        if outcome.state == .confirmedNoChange,
-           outcome.delivery == nil,
-           outcome.dispatchState == .none
-        {
-            return nil
-        }
-        if let failure = DesktopActionFailure(
-            outcome: outcome,
-            message: "Read-only desktop observation returned a non-success or dispatching outcome.",
-            hint: "Observe the target before retrying and update the runtime host.",
-            targetReceipt: targetReceipt)
-        {
-            return failure
-        }
-        return DesktopActionFailure.indeterminate(
-            route: outcome.route,
-            delivery: outcome.delivery,
-            evidence: .completionUnknown,
-            unitCount: outcome.dispatchState.unitCount,
-            message: "Read-only desktop observation contradicted its no-dispatch contract.",
-            hint: "Observe the target before retrying and update the runtime host.")
-            .attributed(to: targetReceipt)
     }
 
     private func handleAutomationRequest(
@@ -1903,5 +1329,3 @@ extension PeekabooBridgeServer {
             && self.windowBoundsProvider(windowID) == capturedBounds
     }
 }
-
-// swiftlint:enable file_length
